Parallel search strategies (#264)
For querying search result items, we try 2 strategies:
- the default strategy is always tried, and is usually the most performant
- for certain searches where items are filtered to a small number of results, and ordered with a limit, the default strategy can be very slow, so we try a CTE strategy, with order and limit on a materialized view of the complete results, and we put it in a race with the default strategy

+++ internal/model/torrents.gen.go
@@ -16,7 +16,7 @@ const TableNameTorrent = "torrents"
 type Torrent struct {
 	InfoHash    protocol.ID             `gorm:"column:info_hash;primaryKey;<-:create" json:"infoHash"`
 	Name        string                  `gorm:"column:name;not null" json:"name"`
-	Size        uint64                  `gorm:"column:size;not null" json:"size"`
+	Size        uint                    `gorm:"column:size;not null" json:"size"`
 	Private     bool                    `gorm:"column:private;not null" json:"private"`
 	CreatedAt   time.Time               `gorm:"column:created_at;not null;<-:create" json:"createdAt"`
 	UpdatedAt   time.Time               `gorm:"column:updated_at;not null" json:"updatedAt"`
